Notification contract and BaseNotification trait
