+++ src/Illuminate/Notifications/Channels/DatabaseChannel.php
@@ -2,7 +2,7 @@
 
 namespace Illuminate\Notifications\Channels;
 
-use Illuminate\Notifications\Notification;
+use Illuminate\Contracts\Notifications\Notification;
 use RuntimeException;
 
 class DatabaseChannel
@@ -11,7 +11,7 @@ class DatabaseChannel
      * Send the given notification.
      *
      * @param  mixed  $notifiable
-     * @param  \Illuminate\Notifications\Notification  $notification
+     * @param  \Illuminate\Contracts\Notifications\Notification  $notification
      * @return \Illuminate\Database\Eloquent\Model
      */
     public function send($notifiable, Notification $notification)
@@ -25,7 +25,7 @@ class DatabaseChannel
      * Get the data for the notification.
      *
      * @param  mixed  $notifiable
-     * @param  \Illuminate\Notifications\Notification  $notification
+     * @param  \Illuminate\Contracts\Notifications\Notification  $notification
      * @return array
      *
      * @throws \RuntimeException
@@ -48,7 +48,7 @@ class DatabaseChannel
      * Build an array payload for the DatabaseNotification Model.
      *
      * @param  mixed  $notifiable
-     * @param  \Illuminate\Notifications\Notification  $notification
+     * @param  \Illuminate\Contracts\Notifications\Notification  $notification
      * @return array
      */
     protected function buildPayload($notifiable, Notification $notification)
