+++ src/Illuminate/Notifications/Events/NotificationFailed.php
@@ -19,7 +19,7 @@ class NotificationFailed
     /**
      * The notification instance.
      *
-     * @var \Illuminate\Notifications\Notification
+     * @var \Illuminate\Contracts\Notifications\Notification
      */
     public $notification;
 
@@ -41,7 +41,7 @@ class NotificationFailed
      * Create a new event instance.
      *
      * @param  mixed  $notifiable
-     * @param  \Illuminate\Notifications\Notification  $notification
+     * @param  \Illuminate\Contracts\Notifications\Notification  $notification
      * @param  string  $channel
      * @param  array  $data
      * @return void
