+++ src/Illuminate/Notifications/Events/NotificationSending.php
@@ -19,7 +19,7 @@ class NotificationSending
     /**
      * The notification instance.
      *
-     * @var \Illuminate\Notifications\Notification
+     * @var \Illuminate\Contracts\Notifications\Notification
      */
     public $notification;
 
@@ -34,7 +34,7 @@ class NotificationSending
      * Create a new event instance.
      *
      * @param  mixed  $notifiable
-     * @param  \Illuminate\Notifications\Notification  $notification
+     * @param  \Illuminate\Contracts\Notifications\Notification  $notification
      * @param  string  $channel
      * @return void
      */
