+++ src/Illuminate/Notifications/Notification.php
@@ -2,46 +2,10 @@
 
 namespace Illuminate\Notifications;
 
+use Illuminate\Contracts\Notifications\Notification as NotificationContract;
 use Illuminate\Queue\SerializesModels;
 
-class Notification
+class Notification implements NotificationContract
 {
-    use SerializesModels;
-
-    /**
-     * The unique identifier for the notification.
-     *
-     * @var string
-     */
-    public $id;
-
-    /**
-     * The locale to be used when sending the notification.
-     *
-     * @var string|null
-     */
-    public $locale;
-
-    /**
-     * Get the channels the event should broadcast on.
-     *
-     * @return array
-     */
-    public function broadcastOn()
-    {
-        return [];
-    }
-
-    /**
-     * Set the locale to send this notification in.
-     *
-     * @param  string  $locale
-     * @return $this
-     */
-    public function locale($locale)
-    {
-        $this->locale = $locale;
-
-        return $this;
-    }
+    use SerializesModels, BaseNotification;
 }
